Rename journal backfill script for consistency
